#88 refactor: TableHook파일 생성
